Added login and register pages for authentication, created store
Implemented login and register pages for the authentication functionality. These pages allow users to log in and create new accounts securely. Additionally, we created the necessary store to manage authentication-related data and states. The store enables efficient management of user authentication data and state changes. With the addition of these pages and the store, we have laid the foundation for a robust and secure authentication system, enhancing the user experience and security of our platform.